requirements.txt
pywin32>=306
python-dotenv>=1.0.0
psutil>=5.9.0